+++ dm-20170622/alibabacloud_dm20170622/models.py
@@ -94,9 +94,6 @@ class AddIpfilterResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -140,6 +137,8 @@ class BatchSendMailRequest(TeaModel):
         resource_owner_id: int = None,
         tag_name: str = None,
         template_name: str = None,
+        un_subscribe_filter_level: str = None,
+        un_subscribe_link_type: str = None,
     ):
         self.account_name = account_name
         self.address_type = address_type
@@ -152,6 +151,8 @@ class BatchSendMailRequest(TeaModel):
         self.resource_owner_id = resource_owner_id
         self.tag_name = tag_name
         self.template_name = template_name
+        self.un_subscribe_filter_level = un_subscribe_filter_level
+        self.un_subscribe_link_type = un_subscribe_link_type
 
     def validate(self):
         pass
@@ -184,6 +185,10 @@ class BatchSendMailRequest(TeaModel):
             result['TagName'] = self.tag_name
         if self.template_name is not None:
             result['TemplateName'] = self.template_name
+        if self.un_subscribe_filter_level is not None:
+            result['UnSubscribeFilterLevel'] = self.un_subscribe_filter_level
+        if self.un_subscribe_link_type is not None:
+            result['UnSubscribeLinkType'] = self.un_subscribe_link_type
         return result
 
     def from_map(self, m: dict = None):
@@ -210,6 +215,10 @@ class BatchSendMailRequest(TeaModel):
             self.tag_name = m.get('TagName')
         if m.get('TemplateName') is not None:
             self.template_name = m.get('TemplateName')
+        if m.get('UnSubscribeFilterLevel') is not None:
+            self.un_subscribe_filter_level = m.get('UnSubscribeFilterLevel')
+        if m.get('UnSubscribeLinkType') is not None:
+            self.un_subscribe_link_type = m.get('UnSubscribeLinkType')
         return self
 
 
@@ -258,9 +267,6 @@ class BatchSendMailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -380,9 +386,6 @@ class CheckDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -508,9 +511,6 @@ class CheckDomainDnsResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -630,9 +630,6 @@ class CreateDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -764,9 +761,6 @@ class CreateMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -898,9 +892,6 @@ class CreateReceiverResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1026,9 +1017,6 @@ class CreateTagResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1196,9 +1184,6 @@ class CreateTemplateResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1312,9 +1297,6 @@ class DeleteDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1428,9 +1410,6 @@ class DeleteMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1544,9 +1523,6 @@ class DeleteReceiverResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1666,9 +1642,6 @@ class DeleteReceiverDetailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1782,9 +1755,6 @@ class DeleteTagResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1904,9 +1874,6 @@ class DeleteTemplateResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2110,9 +2077,6 @@ class DescAccountSummaryResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2400,9 +2364,6 @@ class DescDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2582,9 +2543,6 @@ class DescTemplateResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2828,9 +2786,6 @@ class GetAccountListResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -2956,9 +2911,6 @@ class GetMailAddressMsgCallBackUrlResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3268,9 +3220,6 @@ class GetTrackListResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3396,9 +3345,6 @@ class ModifyMailAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3524,9 +3470,6 @@ class ModifyPWByDomainResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3652,9 +3595,6 @@ class ModifyTagResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -3816,9 +3756,6 @@ class ModifyTemplateResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4098,9 +4035,6 @@ class QueryDomainByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4332,9 +4266,6 @@ class QueryInvalidAddressResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4620,9 +4551,6 @@ class QueryMailAddressByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -4884,9 +4812,6 @@ class QueryReceiverByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5124,9 +5049,6 @@ class QueryReceiverDetailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5352,9 +5274,6 @@ class QueryTagByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5622,9 +5541,6 @@ class QueryTaskByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -5904,9 +5820,6 @@ class QueryTemplateByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6108,9 +6021,6 @@ class SaveReceiverDetailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6354,9 +6264,6 @@ class SenderStatisticsByTagNameAndBatchIDResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6618,9 +6525,6 @@ class SenderStatisticsDetailByParamResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -6668,6 +6572,8 @@ class SingleSendMailRequest(TeaModel):
         tag_name: str = None,
         text_body: str = None,
         to_address: str = None,
+        un_subscribe_filter_level: str = None,
+        un_subscribe_link_type: str = None,
     ):
         self.account_name = account_name
         self.address_type = address_type
@@ -6684,6 +6590,8 @@ class SingleSendMailRequest(TeaModel):
         self.tag_name = tag_name
         self.text_body = text_body
         self.to_address = to_address
+        self.un_subscribe_filter_level = un_subscribe_filter_level
+        self.un_subscribe_link_type = un_subscribe_link_type
 
     def validate(self):
         pass
@@ -6724,6 +6632,10 @@ class SingleSendMailRequest(TeaModel):
             result['TextBody'] = self.text_body
         if self.to_address is not None:
             result['ToAddress'] = self.to_address
+        if self.un_subscribe_filter_level is not None:
+            result['UnSubscribeFilterLevel'] = self.un_subscribe_filter_level
+        if self.un_subscribe_link_type is not None:
+            result['UnSubscribeLinkType'] = self.un_subscribe_link_type
         return result
 
     def from_map(self, m: dict = None):
@@ -6758,6 +6670,10 @@ class SingleSendMailRequest(TeaModel):
             self.text_body = m.get('TextBody')
         if m.get('ToAddress') is not None:
             self.to_address = m.get('ToAddress')
+        if m.get('UnSubscribeFilterLevel') is not None:
+            self.un_subscribe_filter_level = m.get('UnSubscribeFilterLevel')
+        if m.get('UnSubscribeLinkType') is not None:
+            self.un_subscribe_link_type = m.get('UnSubscribeLinkType')
         return self
 
 
@@ -6806,9 +6722,6 @@ class SingleSendMailResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -7158,9 +7071,6 @@ class SingleSendMailV2Response(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()